+++ argopy/xarray.py
@@ -24,6 +24,7 @@ from .utils.compute import (
     groupby_remap,
 )
 from .utils.geo import toYearFraction
+from .utils.decorators import deprecated
 from .errors import InvalidDatasetStructure, DataNotFound, OptionValueError
 
 
@@ -576,6 +577,11 @@ class ArgoAccessor:
         ds.argo._type = "point"
         return ds
 
+    @deprecated(
+        "This method is deprecated and will break your code for versions >= 0.1.18. ",
+        ignore_caller="postprocessing",
+        version="0.1.17",
+    )
     def filter_data_mode(  # noqa: C901
         self, keep_error: bool = True, errors: str = "raise"
     ):
@@ -602,6 +608,16 @@ class ArgoAccessor:
         Returns
         -------
         :class:`xarray.Dataset`
+
+
+        .. deprecated:: 0.1.7
+
+            This method action and signature are deprecated and will break your code for versions >= 0.1.18.
+
+            In versions >= 0.1.18, this method behavior will be transferred to the new `transform_data_mode` method
+            while this method `filter_data_mode` will change its signature and behavior to really filter measurements
+            according to DATA_MODE or <PARAM>_DATA_MODE values.
+
         """
         if self._type != "point":
             raise InvalidDatasetStructure(