+++ docs/whats-new.rst
@@ -11,13 +11,26 @@ What's New
 Coming up next
 --------------
 
+Features and front-end API
+^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+.. important:: Introduces deprecation warnings before the upcoming major release v1.0.0. (:pr:`389`) by `G. Maze <http://www.github.com/gmaze>`_.
+
+    .. currentmodule:: xarray
+
+    - Change of signature and action with xarray Argo accessor :meth:`Dataset.argo.filter_data_mode`
+
+    .. currentmodule:: argopy
+
+    - Refactor option "dataset" into "ds", see :class:`argopy.set_options`
+    - Refactor option "ftp" into "gdac", see :class:`argopy.set_options`
+
 Internals
 ^^^^^^^^^
 
 - Refactor Argovis CI tests to use mocked http server (:pr:`383`) by `G. Maze <http://www.github.com/gmaze>`_
 
 - Improve error and warning messages from mocked http server to address :issue:`381` (:pr:`382`) by `G. Maze <http://www.github.com/gmaze>`_
-
 
 
 v0.1.16 (27 Aug. 2024)
@@ -71,6 +84,7 @@ Breaking changes
 
 - Drop support for erddapy < v0.8.0 (:pr:`344`) by `G. Maze <http://www.github.com/gmaze>`_.
 
+
 v0.1.15 (12 Dec. 2023)
 ----------------------
 
@@ -78,6 +92,7 @@ Internals
 ^^^^^^^^^
 
 - Fix bug whereby user name could not be retrieved using :func:`getpass.getuser`. This closes :issue:`310` and allows argopy to be integrated into the EU Galaxy tools for `ecology <https://github.com/galaxyecology/tools-ecology/pull/81>`_. (:pr:`311`) by `G. Maze <http://www.github.com/gmaze>`_.
+
 
 v0.1.14 (29 Sep. 2023)
 ----------------------
@@ -335,6 +350,7 @@ Breaking changes
 ^^^^^^^^^^^^^^^^
 
 - Some documentation pages may have moved to new urls.
+
 
 v0.1.14rc1 (31 May 2023)
 ------------------------
@@ -622,6 +638,7 @@ Internals
 
 - Update and clean up requirements. Remove upper bound on all dependencies (:pr:`182`) by `R. Abernathey <http://www.github.com/rabernat>`_.
 
+
 v0.1.9 (19 Jan. 2022)
 ---------------------
 
@@ -791,6 +808,7 @@ Internals
         handlers=[logging.FileHandler("argopy.log", mode='w')]
     )
 
+
 v0.1.7 (4 Jan. 2021)
 -----------------------
 
@@ -927,6 +945,7 @@ Internals
 
 - Erddap fetcher now uses netcdf format to retrieve data (:pr:`19`).
 
+
 v0.1.3 (15 May 2020)
 --------------------
 
@@ -976,10 +995,12 @@ Internals
 
 - Improved unit testing (:commit:`e9555d1e6e90d3d1e75183cec0c4e14f7f19c17c`, :commit:`4b60ede844e37df86b32e4e2a2008335472a8cc1`, :commit:`34abf4913cb8bec027f88301c5504ebe594b3eae`)
 
+
 v0.1.2 (15 May 2020)
 --------------------
 
 We didn't like this one this morning, so we move one to the next one !
+
 
 v0.1.1 (3 Apr. 2020)
 ---------------------